cursedwand.java
- largely overwrote with remote version, but added checks for user missing whenever possible
- need to make sure chaos trap doesn't crash the game over and over again lmao
- probably just need to make sure ballistica is subbed in when needed.

Wand.java:
- lingering magic needs to be buffed
- added kings wisdom to be tied to lingering magic
- removed overcharge mechanic, we will see if we can actually get away with it.

Actor:
- Did I accidentally miss curActorPriority to be implemented? It's implemented now.

Dungeon.java:
- minor conflict regarding reveal method, but it was just independent logic re mimics being added.

Char:
- attempted to merge the damage changes

Belongings:
- implemented shard of oblivion

Chasm:
- implemented visuals that were missing due to a merge mistake

WndInfoTalent
- Largely ignored incoming changes, as they seem to be redundant with me making it inherit from titled message

DK 584: ignored adding of an invulnerable check

items.png: moved crown by 1, moved broken crown by 1, moved trinket by 1

ItemSlot.java: ClassArmor opts out of the item icon changes for now

need to add null check for crossbow.find

games in progress:
- moved incompatibility to 1.4.3

actors.properties:
be sure to implement as intro_quest_start

talents:
- focused meal nerfed to 1/1.5 instead of .67/1, will probably buff
- identify vfx still plays with shard of oblivion, but does not actually identify
- implemented removal of on-id talents
- hearty meal implemented, but need to implement unique hearty meal effect
- monastic might now updates quickslot
- todo check if shard of oblivion works properly with instant id talents
- adjusted lethal momentum / lethal rush implementations slightly:
   - lethal rush now gives the standard shpd effect, but it also gives adrenaline

HornOfPlenty had a weird merge conflict, will need to test it.

GameScene: overwrote rogue's foresight, will need to re-implement

Item.java: make talent.onitemcollected runs catalog id when appropriate

RevealedArea:
- refactored to not even ask for depth anymore

actors.properties
- imp greeting->intro

kindofweapon
- merged doEquip

talents to re-implement:
* hearty meal
* cached rations
* focused meal
* cached rations (sort of)
* liquid agility

dagger
- overwrote my sneak with their sneak

mace
- heavy blow is now an inner class, has infinite accuracy, adjusted damage multi

MeleeWeapon:
- 11a410f0 removed alternate chargers - implemented
- removed `usesTargeting = false` from MeleeWeapon
- aggressive barrier needs to be re-implemented
- need to implement lethal haste changes in Talent.java
- was precise assault implementation moved? also need to reimplement twin upgrades
- added a delayMulti instance variable to MeleeAbility for Sword

MissileWeapon:
- charges shown for duelist are now just the regular charges
- fixed level adjustment from rogue cloak not affecting durability

magesstaff:
- overwrote wand preservation

Sniper's Mark:
- he overwrote level with percentDamageBonus I guess??? kept both for now, but added the new value as a required value
- balance changed from 10/20/30/40 flat to 2.5/5/10/15, which is a huge nerf to +0 (though this is just going to end up being standardized) and a -50%/-33%/-25% to nerf to shurikens in particular. compared to shattered, though, it's still kinda buffed.

ShieldBuff:
- removed redundant vfx calls. there might be a few *extra* vfx calls out in the wild if Evan added a usage without the corresponding vfx call.

armorability:
- implemented heroic energy representation change, but it's still the same values

MonkEnergy.java:
- attempt to implement f7c6a2ad289891ed96f8718fd5b50b92a09a8091

HoldFast.java:
- removed HoldFast#armor, replaced with HoldFast.armorMin and HoldFast.armorMax

Imp.java: added a hero name argument 00b4300f0f46fd0e511024c5405bfc7478ddbe16

rings:
- evan adjusted some numbers, but I don't have to change as many numbers because I can just change Ring#multiplier() for a lot of them. I don't know how he sleeps at night with all that duplication. It's literally all I can do to not refactor immediately.

TitleScene:
- blindly adjusted

+++ core/src/main/java/com/zrp200/rkpd2/items/weapon/missiles/HeavyBoomerang.java
@@ -66,7 +66,7 @@ public class HeavyBoomerang extends MissileWeapon {
 		}
 		decrementDurability();
 		if (durability > 0){
-			Buff.append(Dungeon.hero, CircleBack.class).setup(this, cell, Dungeon.hero.pos, Dungeon.depth);
+			Buff.append(Dungeon.hero, CircleBack.class).setup(this, cell, Dungeon.hero.pos, Dungeon.depth, Dungeon.branch);
 		}
 	}
 
@@ -77,7 +77,7 @@ public class HeavyBoomerang extends MissileWeapon {
 			return;
 		}
 		parent = null;
-		Buff.append(Dungeon.hero, CircleBack.class).setup(this, cell, Dungeon.hero.pos, Dungeon.depth);
+		Buff.append(Dungeon.hero, CircleBack.class).setup(this, cell, Dungeon.hero.pos, Dungeon.depth, Dungeon.branch);
 	}
 	
 	public static class CircleBack extends Buff {
@@ -90,14 +90,16 @@ public class HeavyBoomerang extends MissileWeapon {
 		private int thrownPos;
 		private int returnPos;
 		private int returnDepth;
-		
+		private int returnBranch;
+
 		private int left;
 		
-		public void setup( HeavyBoomerang boomerang, int thrownPos, int returnPos, int returnDepth){
+		public void setup( HeavyBoomerang boomerang, int thrownPos, int returnPos, int returnDepth, int returnBranch){
 			this.boomerang = boomerang;
 			this.thrownPos = thrownPos;
 			this.returnPos = returnPos;
 			this.returnDepth = returnDepth;
+			this.returnBranch = returnBranch;
 			left = 3;
 		}
 		
@@ -116,7 +118,7 @@ public class HeavyBoomerang extends MissileWeapon {
 		
 		@Override
 		public boolean act() {
-			if (returnDepth == Dungeon.depth){
+			if (returnDepth == Dungeon.depth && returnBranch == Dungeon.branch){
 				left--;
 				if (left <= 0){
 					final Char returnTarget = Actor.findChar(returnPos);
@@ -127,19 +129,22 @@ public class HeavyBoomerang extends MissileWeapon {
 									boomerang,
 							() -> {
 								if (returnTarget == target){
-									if (target instanceof Hero && boomerang.doPickUp((Hero) target)) {
-										//grabbing the boomerang takes no time
-										((Hero) target).spend(-TIME_TO_PICK_UP);
-									} else {
-										Dungeon.level.drop(boomerang, returnPos).sprite.drop();
-									}
-
+                                    if (!boomerang.spawnedForEffect) {
+                                        if (target instanceof Hero && boomerang.doPickUp((Hero) target)) {
+                                            //grabbing the boomerang takes no time
+                                            ((Hero) target).spend(-TIME_TO_PICK_UP);
+                                        } else {
+                                            Dungeon.level.drop(boomerang, returnPos).sprite.drop();
+                                        }
+                                    }
 								} else if (returnTarget != null){
 									boomerang.circleBackHit = true;
 									((Hero)target).shoot( returnTarget, boomerang );
 									boomerang.circleBackHit = false;
 								} else {
-									Dungeon.level.drop(boomerang, returnPos).sprite.drop();
+                                    if (!boomerang.spawnedForEffect) {
+                                        Dungeon.level.drop(boomerang, returnPos).sprite.drop();
+                                    }
 								}
 								CircleBack.this.next();
 							});
@@ -158,7 +163,8 @@ public class HeavyBoomerang extends MissileWeapon {
 		private static final String THROWN_POS = "thrown_pos";
 		private static final String RETURN_POS = "return_pos";
 		private static final String RETURN_DEPTH = "return_depth";
-		
+		private static final String RETURN_BRANCH = "return_branch";
+
 		@Override
 		public void storeInBundle(Bundle bundle) {
 			super.storeInBundle(bundle);
@@ -166,6 +172,7 @@ public class HeavyBoomerang extends MissileWeapon {
 			bundle.put(THROWN_POS, thrownPos);
 			bundle.put(RETURN_POS, returnPos);
 			bundle.put(RETURN_DEPTH, returnDepth);
+			bundle.put(RETURN_BRANCH, returnBranch);
 		}
 		
 		@Override
@@ -175,6 +182,7 @@ public class HeavyBoomerang extends MissileWeapon {
 			thrownPos = bundle.getInt(THROWN_POS);
 			returnPos = bundle.getInt(RETURN_POS);
 			returnDepth = bundle.getInt(RETURN_DEPTH);
+			returnBranch = bundle.contains(RETURN_BRANCH) ? bundle.getInt(RETURN_BRANCH) : 0;
 		}
 	}
