+++ core/src/main/java/com/zrp200/rkpd2/items/weapon/missiles/darts/TippedDart.java
@@ -32,8 +32,22 @@ import com.zrp200.rkpd2.items.Generator;
 import com.zrp200.rkpd2.items.Item;
 import com.zrp200.rkpd2.items.wands.WandOfRegrowth;
 import com.zrp200.rkpd2.items.weapon.melee.Crossbow;
+import com.zrp200.rkpd2.journal.Catalog;
 import com.zrp200.rkpd2.messages.Messages;
-import com.zrp200.rkpd2.plants.*;
+import com.zrp200.rkpd2.plants.Blindweed;
+import com.zrp200.rkpd2.plants.Dreamfoil;
+import com.zrp200.rkpd2.plants.Earthroot;
+import com.zrp200.rkpd2.plants.Fadeleaf;
+import com.zrp200.rkpd2.plants.Firebloom;
+import com.zrp200.rkpd2.plants.Icecap;
+import com.zrp200.rkpd2.plants.Mageroyal;
+import com.zrp200.rkpd2.plants.Plant;
+import com.zrp200.rkpd2.plants.Rotberry;
+import com.zrp200.rkpd2.plants.Sorrowmoss;
+import com.zrp200.rkpd2.plants.Starflower;
+import com.zrp200.rkpd2.plants.Stormvine;
+import com.zrp200.rkpd2.plants.Sungrass;
+import com.zrp200.rkpd2.plants.Swiftthistle;
 import com.zrp200.rkpd2.scenes.GameScene;
 import com.zrp200.rkpd2.sprites.ItemSprite;
 import com.zrp200.rkpd2.windows.WndOptions;
@@ -41,6 +55,7 @@ import com.watabou.utils.Reflection;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 
 public abstract class TippedDart extends Dart {
 	
@@ -64,13 +79,25 @@ public abstract class TippedDart extends Dart {
 	public void execute(final Hero hero, String action) {
 		super.execute(hero, action);
 		if (action.equals( AC_CLEAN )){
-			
+
+			String[] options;
+			if (quantity() > 1){
+				options = new String[]{
+					Messages.get(this, "clean_all"),
+					Messages.get(this, "clean_one"),
+					Messages.get(this, "cancel")
+				};
+			} else {
+				options = new String[]{
+					Messages.get(this, "clean_one"),
+					Messages.get(this, "cancel")
+				};
+			}
+
 			GameScene.show(new WndOptions(new ItemSprite(this),
 					Messages.titleCase(name()),
 					Messages.get(this, "clean_desc"),
-					Messages.get(this, "clean_all"),
-					Messages.get(this, "clean_one"),
-					Messages.get(this, "cancel")){
+					options){
 				@Override
 				protected void onSelect(int index) {
 					if (index == 0){
@@ -80,7 +107,7 @@ public abstract class TippedDart extends Dart {
 						hero.spend( 1f );
 						hero.busy();
 						hero.sprite.operate(hero.pos);
-					} else if (index == 1){
+					} else if (index == 1 && quantity() > 1){
 						detach(hero.belongings.backpack);
 						if (!new Dart().collect()) Dungeon.level.drop(new Dart(), hero.pos).sprite.drop();
 
@@ -105,9 +132,10 @@ public abstract class TippedDart extends Dart {
 		super.rangedHit( enemy, cell);
 		
 		//need to spawn a dart
-		if (durability <= 0){
+		if (durability <= 0 && !spawnedForEffect){
 			//attempt to stick the dart to the enemy, just drop it if we can't.
 			Dart d = new Dart();
+			Catalog.countUse(getClass());
 			if (sticky && enemy != null && enemy.isAlive() && enemy.alignment != Char.Alignment.ALLY){
 				PinCushion p = Buff.affect(enemy, PinCushion.class);
 				if (p.target == enemy){
@@ -140,42 +168,46 @@ public abstract class TippedDart extends Dart {
 	@Override
 	public float durabilityPerUse() {
 		float use = super.durabilityPerUse(false);
-
-		int points = Dungeon.hero.pointsInTalent(Talent.DURABLE_TIPS, Talent.RK_WARDEN);
+if (Dungeon.hero != null) {
+			int points = Dungeon.hero.pointsInTalent(Talent.DURABLE_TIPS, Talent.RK_WARDEN);
 		if(Dungeon.hero.canHaveTalent(Talent.DURABLE_TIPS)) points++;
 		use /= (1 + points);
 
-		//checks both destination and source position
-		float lotusPreserve = 0f;
-		if (targetPos != -1){
-			for (Char ch : Actor.chars()){
-				if (ch instanceof WandOfRegrowth.Lotus){
+			//checks both destination and source position
+			float lotusPreserve = 0f;
+			if (targetPos != -1) {
+				for (Char ch : Actor.chars()) {
+					if (ch instanceof WandOfRegrowth.Lotus) {
+						WandOfRegrowth.Lotus l = (WandOfRegrowth.Lotus) ch;
+						if (l.inRange(targetPos)) {
+							lotusPreserve = Math.max(lotusPreserve, l.seedPreservation());
+						}
+					}
+				}
+				targetPos = -1;
+			}
+			int p = curUser == null ? Dungeon.hero.pos : curUser.pos;
+			for (Char ch : Actor.chars()) {
+				if (ch instanceof WandOfRegrowth.Lotus) {
 					WandOfRegrowth.Lotus l = (WandOfRegrowth.Lotus) ch;
-					if (l.inRange(targetPos)){
+					if (l.inRange(p)) {
 						lotusPreserve = Math.max(lotusPreserve, l.seedPreservation());
 					}
 				}
 			}
-			targetPos = -1;
+			use *= (1f - lotusPreserve);
 		}
-		int p = curUser == null ? Dungeon.hero.pos : curUser.pos;
-		for (Char ch : Actor.chars()){
-			if (ch instanceof WandOfRegrowth.Lotus){
-				WandOfRegrowth.Lotus l = (WandOfRegrowth.Lotus) ch;
-				if (l.inRange(p)){
-					lotusPreserve = Math.max(lotusPreserve, l.seedPreservation());
-				}
-			}
-		}
-		use *= (1f - lotusPreserve);
 
 		float usages = Math.round(MAX_DURABILITY/use);
 
-		//grants 4 extra uses with charged shot
-		if (Dungeon.hero.buff(Crossbow.ChargedShot.class) != null){
-			usages += 4;
+		//grants 3+lvl extra uses with charged shot
+		if (bow != null && Dungeon.hero != null && Dungeon.hero.buff(Crossbow.ChargedShot.class) != null){
+			usages += 3 + bow.buffedLvl();
 		}
-//add a tiny amount to account for rounding error for calculations like 1/3
+//at 100 uses, items just last forever.
+		if (usages >= 100f) return 0;
+
+		//add a tiny amount to account for rounding error for calculations like 1/3
 		return (MAX_DURABILITY/usages) + 0.001f;
 	}
 	
@@ -185,21 +217,22 @@ public abstract class TippedDart extends Dart {
 		return 8 * quantity;
 	}
 	
-	private static HashMap<Class<?extends Plant.Seed>, Class<?extends TippedDart>> types = new HashMap<>();
+	public static final LinkedHashMap<Class<?extends Plant.Seed>, Class<?extends TippedDart>> types = new LinkedHashMap<>();
 	static {
-		types.put(Blindweed.Seed.class,     BlindingDart.class);
-		types.put(Dreamfoil.Seed.class,     DreamDart.class);
-		types.put(Mageroyal.Seed.class,		CleansingDart.class);
-		types.put(Earthroot.Seed.class,     ParalyticDart.class);
-		types.put(Fadeleaf.Seed.class,      DisplacingDart.class);
-		types.put(Firebloom.Seed.class,     IncendiaryDart.class);
-		types.put(Icecap.Seed.class,        ChillingDart.class);
-		types.put(Rotberry.Seed.class,      RotDart.class);
-		types.put(Sorrowmoss.Seed.class,    PoisonDart.class);
-		types.put(Starflower.Seed.class,    HolyDart.class);
-		types.put(Stormvine.Seed.class,     ShockingDart.class);
+        types.put(Dreamfoil.Seed.class,     DreamDart.class);
+
+        types.put(Rotberry.Seed.class,      RotDart.class);
 		types.put(Sungrass.Seed.class,      HealingDart.class);
+		types.put(Fadeleaf.Seed.class,      DisplacingDart.class);
+		types.put(Icecap.Seed.class,        ChillingDart.class);
+		types.put(Firebloom.Seed.class,     IncendiaryDart.class);
+		types.put(Sorrowmoss.Seed.class,    PoisonDart.class);
 		types.put(Swiftthistle.Seed.class,  AdrenalineDart.class);
+		types.put(Blindweed.Seed.class,     BlindingDart.class);
+		types.put(Stormvine.Seed.class,     ShockingDart.class);
+		types.put(Earthroot.Seed.class,     ParalyticDart.class);
+		types.put(Mageroyal.Seed.class,     CleansingDart.class);
+		types.put(Starflower.Seed.class,    HolyDart.class);
 	}
 	
 	public static TippedDart getTipped( Plant.Seed s, int quantity ){
